Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"django>=6.0.0,<7.0.0",
	"djangorestframework>=3.17.0,<4.0.0",
	"django-cors-headers>=4.9.0,<5.0.0",
]
